Cat pillow update!
added a couple bonus features including cat pillow and the ability to turn off the popups and screen effects, as well as cleaned some stuff up. There is a title screen now to initiate the music, because audio refuses to play without a screen event. Also more buttons on the top left, that seems like it's becoming my thing a little and I like it